+++ src/Controller/LoginController.php
@@ -14,6 +14,11 @@
 namespace Eventum\Controller;
 
 use Auth;
+use AuthCookie;
+use Eventum\Auth\Adapter\AdapterInterface;
+use Eventum\Auth\AuthException;
+use Eventum\Session;
+use User;
 use Validation;
 
 class LoginController extends BaseController
@@ -22,27 +27,30 @@ class LoginController extends BaseController
     private $login;
     /** @var string */
     private $passwd;
-
     /** @var string */
     private $url;
+    /** @var bool */
+    private $remember;
+    /** @var AdapterInterface */
+    private $auth;
 
     /**
      * {@inheritdoc}
      */
-    protected function configure()
+    protected function configure(): void
     {
         $post = $this->getRequest()->request;
 
-        $this->login = (string) $post->get('email');
-        $this->passwd = (string) $post->get('passwd');
-
-        $this->url = (string) $post->get('url');
+        $this->login = (string)$post->get('email');
+        $this->passwd = (string)$post->get('passwd');
+        $this->remember = (bool)$post->get('remember');
+        $this->url = (string)$post->get('url');
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function canAccess()
+    protected function canAccess(): bool
     {
         return true;
     }
@@ -50,33 +58,15 @@ class LoginController extends BaseController
     /**
      * {@inheritdoc}
      */
-    protected function defaultAction()
+    protected function defaultAction(): void
     {
-        if (Validation::isWhitespace($this->login)) {
-            $this->redirect('index.php?err=1');
+        $this->auth = Auth::getAuthBackend();
+        try {
+            $this->authenticate($this->login, $this->passwd);
+            $this->login($this->login, $this->remember);
+        } catch (AuthException $e) {
+            $this->loginFailure($e->getCode(), $e->getMessage(), ['login' => $this->login]);
         }
-
-        if (Validation::isWhitespace($this->passwd)) {
-            $this->loginFailure(2, 'empty password', ['email' => $this->login]);
-        }
-
-        // check if user exists
-        if (!Auth::userExists($this->login)) {
-            $this->loginFailure(3, 'unknown user');
-        }
-
-        // check if user is locked
-        $usr_id = Auth::getUserIDByLogin($this->login);
-        if (Auth::isUserBackOffLocked($usr_id)) {
-            $this->loginFailure(13, 'account back-off locked');
-        }
-
-        // check if the password matches
-        if (!Auth::isCorrectPassword($this->login, $this->passwd)) {
-            $this->loginFailure(3, 'wrong password', ['email' => $this->login]);
-        }
-
-        Auth::login($this->login);
 
         $params = [];
         if ($this->url) {
@@ -86,6 +76,46 @@ class LoginController extends BaseController
         $this->redirect('select_project.php', $params);
     }
 
+    private function authenticate(string $login, string $passwd): void
+    {
+        if (Validation::isWhitespace($login)) {
+            throw new AuthException('empty login', AuthException::EMPTY_LOGIN);
+        }
+
+        if (Validation::isWhitespace($passwd)) {
+            throw new AuthException('empty password', AuthException::EMPTY_PASSWORD);
+        }
+
+        if (!$this->auth->userExists($login)) {
+            throw new AuthException('unknown user', AuthException::UNKNOWN_USER);
+        }
+
+        if (!$this->auth->verifyPassword($login, $passwd)) {
+            throw new AuthException('wrong password', AuthException::WRONG_PASSWORD);
+        }
+    }
+
+    private function login(string $login, bool $remember): void
+    {
+        // get user primary mail,
+        // handle aliases since the user is now authenticated
+        $usr_id = Auth::getUserIDByLogin($login);
+        $login = User::getEmail($usr_id);
+
+        // check if this user did already confirm his account
+        if (Auth::isPendingUser($login)) {
+            throw new AuthException('pending user', AuthException::PENDING_USER);
+        }
+
+        if (!Auth::isActiveUser($login)) {
+            throw new AuthException('inactive user', AuthException::INACTIVE_USER);
+        }
+
+        Auth::saveLoginAttempt($login, 'success');
+        AuthCookie::setAuthCookie($login, $remember);
+        Session::init($usr_id);
+    }
+
     /**
      * Log login failure and redirect to login form
      *
@@ -93,7 +123,7 @@ class LoginController extends BaseController
      * @param string $reason
      * @param array $params
      */
-    private function loginFailure($error, $reason, $params = [])
+    private function loginFailure($error, $reason, $params = []): void
     {
         Auth::saveLoginAttempt($this->login, 'failure', $reason);