+++ CHANGELOG.md
@@ -16,6 +16,7 @@ first, also the PHP version required is 7.1, see [#426].
 - TextMessage: do not parse subparts if no content-type given (@glensc, #452)
 - Allow auth adapters to be chained (@glensc, #441)
 - Remove Legacy CLI application (@glensc, #453)
+- Move Login-Back-Off logic to MySQL adapter only (@glensc, #455)
 
 [3.6.0]: https://github.com/eventum/eventum/compare/v3.5.6...master
 [#426]: https://github.com/eventum/eventum/pull/426